refactor: Extract JavaScript prune rendering pure functions
Move manifest/workspace/patch/bin prune helpers into prune_js without
behavior changes so orchestration can migrate onto knowledge next.

Closes TURBO-5833.
